untangle: pinning moves peer/transport/ -> access/ (pure move)
Fingerprint parse/format/verifier vocabulary is cert-layer material:
access/certs.rs and cert_server.rs consumed it from peer/transport,
the remaining wrong-direction refs of the access->peer edge. File
moves verbatim; transport/mod.rs re-exports the module at the old
mount (its item re-exports chain through), so all ten consumer files
keep compiling; access-internal refs re-anchor to
crate::access::pinning.
