+++ src/bin/caller/access/mod.rs
@@ -26,6 +26,7 @@ pub mod client_key;
 pub mod enrollment;
 pub mod iam;
 pub mod org;
+pub mod pinning;
 pub mod state;
 #[cfg(not(target_os = "windows"))]
 pub mod wizard;
